backend/requirements.txt
fastapi
uvicorn
sqlmodel
mysql-connector-python
python-multipart
python-dotenv
pdfplumber
deep-translator
